requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.2
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
lxml==6.0.2
multidict==6.6.4
numpy==2.3.3
packaging==25.0
pandas==2.3.3
praw==7.8.1
prawcore==2.4.0
propcache==0.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.8.0
yarl==1.20.1
